Refactoring steps classes into their own package
