+++ src/test/java/org/eaa690/aerie/steps/RosterWebSteps.java
@@ -14,10 +14,11 @@
  *  limitations under the License.
  */
 
-package org.eaa690.aerie;
+package org.eaa690.aerie.steps;
 
 import io.cucumber.java.en.When;
 import io.restassured.http.ContentType;
+import org.eaa690.aerie.TestContext;
 
 /**
  * Roster Web test steps.
